Add protos to be used in tests.
Also, add a script for regenerating the protos using the latest Go
protobuf library.

The generated Go code for the protos is checked in the repo because it
will be useful for reproducing test results and will also be useful for
comparing the generated code in the event that an update changes it.