Add support for hashing `google.protobuf.Timestamp` protos.
This largely follows the current structure of the Timestamp proto.
Essentially, a `Timestamp` is considered equivalent to a list of two
integers, where:
  - The first list item is a non-negative integer equal to the
    timestamp's UTC seconds since epoch.
  - The second list item is a non-negative integer equal to the
    timestamp's fractions of a second at nanosecond resolution.

Additionally, the semantics of the Timestamp object imply that the
distinction between unset and zero happen at the message level, rather
than the field level.  As a result, an unset timestamp is one where the
proto itself is nil, while an explicitly set timestamp with unset fields
is considered to be explicitly set to 0.  This is unlike normal proto3
messages, where unset/zero fields must be considered to be unset,
because they're indistinguishable in the general case.

Note: If necessary, future versions of ObjectHash-Proto might introduce
options to control how those `Timestamp`s are hashed. However, this is
not currently needed.