Improve the errors returned by the detection module.
This is because the currently returned errors are misleading and
confusing.

Before this change, when compiling protos using the latest
protoc-gen-go, the following error would be returned when attempting to
hash a proto:

```
got an unexpected struct type: reflect.Value
```

With this change, the error message would be similar to the following:

```
got an unexpected struct of type 'struct {}' for field {Name:XXX_NoUnkeyedLiteral PkgPath: Type:struct {} Tag:json:"-" Offset:96 Index:[7] Anonymous:false}
```

The tests added in this change check that the errors returned match what
is expected. This is to prevent regressions in the quality of those
error messages.